# By wangli (5) and others
# Via wangli (2) and liliang (1)
* 'master' of https://github.com/nicolastinkl/AISBOSS:
  调试增加心愿单接口
  增加心愿单笔记
  完善参数修改回调
  增加参数上传接口
  优化参数保存
  Delete hehe.strings
  增加 service deleted status 枚举
  修改数据模型

+++ AISBOSS/ViewController/Buyer/AIBubble.h
@@ -28,7 +28,7 @@ typedef enum  {
 @property (nonatomic) BOOL hadRecommend;
 
 @property (nonatomic) CGFloat radius;
-@property (nonatomic) int index;
+@property (nonatomic) NSInteger index;
 
 //类型
 @property (nonatomic) BubbleType bubbleType;
@@ -64,7 +64,7 @@ typedef enum  {
 
 - (instancetype)initWithFrame:(CGRect)frame model:(AIBuyerBubbleModel *)model;
 
-- (instancetype)initWithCenter:(CGPoint)center model:(AIBuyerBubbleModel *)model type:(BubbleType) type Index:(int) indexModel;
+- (instancetype)initWithCenter:(CGPoint)center model:(AIBuyerBubbleModel *)model type:(BubbleType) type Index:(NSInteger) indexModel;
 
 
 @end
